refactor: remove obsolete code and rename classes
